Create NumberOfSteps.java
/* You are given two arrays and a1,a2...an and b1,b2...bn. In each step, you can set if ai = ai - bi, if ai>= bi.
Determine the minimum number of steps that are required to make all a's equal.
Input format:
First line: n
Second line: a1,a2..an 
Third line: b1,b2..bn

Output format:
Print the minimum number of steps that are required to make all 
's equal. If it is not possible, then print -1. */